+++ views/js/admin/library/src/shared/types/index.ts
@@ -105,6 +105,10 @@ export interface MolliePaymentMethodsTranslations {
   applePayButtonOutline: string;
   applePayButtonWhite: string;
   applePayDirectCertificateConflict: string;
+  applePayExcludedCarriers: string;
+  selectCarriersToExclude: string;
+  applePayExcludedCarriersHelp: string;
+  applePayAllCarriersExcludedWarning: string;
   ignore: string;
 
   // Payment Restrictions